third_party/blink/web_tests/platform/generic/fast/events/domactivate-sets-underlying-click-event-as-handled-expected.txt

To test manually, click on all the form elements (except the textbox) and the details/summary element. The anchor shouldn't be activated (i.e. you shouldn't see any "anchor activated" messages).

Focusing [object HTMLInputElement] type=text and pressing enter
form submitted

Activated [object HTMLInputElement] type=submit
Focusing [object HTMLInputElement] type=submit and pressing enter
Activated [object HTMLInputElement] type=submit
form submitted

form submitted

Activated [object HTMLInputElement] type=image
Focusing [object HTMLInputElement] type=image and pressing enter
Activated [object HTMLInputElement] type=image
form submitted

form submitted

Activated [object HTMLButtonElement] type=submit
Focusing [object HTMLButtonElement] type=submit and pressing enter
Activated [object HTMLButtonElement] type=submit
form submitted

form submitted

Activated [object HTMLInputElement] type=reset
Focusing [object HTMLInputElement] type=reset and pressing enter
Activated [object HTMLInputElement] type=reset
Activated [object HTMLButtonElement] type=reset
Focusing [object HTMLButtonElement] type=reset and pressing enter
Activated [object HTMLButtonElement] type=reset
Activated [object HTMLInputElement] type=checkbox
Activated [object HTMLInputElement] type=radio
Activated [object HTMLInputElement] type=file
Focusing [object HTMLInputElement] type=file and pressing enter
Activated [object HTMLInputElement] type=file

The details element was opened
